Adding ability to delete a IIIF collection
